WIP: saving

+++ src/Logic/Hacks/LockSeed.cpp
@@ -2,7 +2,7 @@
 using namespace geode::prelude;
 
 bool seedHackEnabled = false;
-int seedValue = 1;
+float seedValue = 1.0f;
 class $modify(GameToolbox) {
     #ifdef GEODE_IS_WINDOWS 
     // windows code here
@@ -12,7 +12,7 @@ class $modify(GameToolbox) {
             return GameToolbox::fast_rand_0_1();   // keeps old seed if enabled
 
 
-        int seed = static_cast<int>(seedValue);
+        int seed = static_cast<float>(seedValue);
         int newSeed = (214013 * seed) + 2531011;
         return ((newSeed >> 16) & 0x7FFF) / 32767.f;
     }
